% --- Source file: ex_types.Rd ---
\name{ex_types}
\alias{ex_types}
%\alias{N00}
%\alias{N01}
%\alias{N10}
%\alias{N11}

%\docType{package}
\title{
  Data for the h.types example
}
\description{
 Sample data for \code{\link{h.types}}
}
\details{ The data frame contains columns for disease subtype, study center, and 3 SNPs.
}
\examples{

 data(ex_types, package="ASSET")

 # Display the first 10 rows of the data
 data[1:10, ]

}
\keyword{data}
\seealso{h.traits}



